sqlite executeUpdate 返回值已正常

+++ include/sese/db/DriverInstance.h
@@ -10,6 +10,6 @@ namespace sese::db {
 
         virtual ~DriverInstance() noexcept = default;
         virtual ResultSet::Ptr executeQuery(const char *sql) const noexcept = 0;
-        virtual uint64_t executeUpdate(const char *sql) const noexcept = 0;
+        virtual size_t executeUpdate(const char *sql) const noexcept = 0;
     };
 }// namespace sese::db
